+++ tests/test-functions.php
@@ -12,7 +12,7 @@ class TestFunctions extends TimmyUnitTestCase {
 
 		$image = $this->get_upload_url() . '/test-1400x0-c-default.jpg';
 
-		$this->assertEquals( $result, $image );
+		$this->assertEquals( $image, $result );
 	}
 
 	public function test_get_timber_image() {
@@ -21,7 +21,67 @@ class TestFunctions extends TimmyUnitTestCase {
 
 		$image = ' src="' . $this->get_upload_url() . '/test-1400x0-c-default.jpg" alt=""';
 
-		$this->assertEquals( $result, $image );
+		$this->assertEquals( $image, $result );
+	}
+
+	/**
+	 * Tests whether we get a scaled or a non-scaled image in return.
+	 *
+	 * With images bigger than 2560px, WordPress will create an image version with '-scaled' in its
+	 * name. We always want the image to be created from the original size.
+	 *
+	 * @since 0.14.4
+	 */
+	public function test_get_timber_image_huge_nonscaled() {
+		$attachment = $this->create_image( [ 'file' => 'huge.jpg' ] );
+		$result     = get_timber_image( $attachment, 'large' );
+
+		$image = ' src="' . $this->get_upload_url() . '/huge-1400x0-c-default.jpg" alt=""';
+
+		$this->assertEquals( $image, $result );
+	}
+
+	/**
+	 * Tests whether a '-scaled' version of the image is returned when the
+	 * 'big_image_size_threshold' kicks in.
+	 */
+	public function test_get_timber_image_full_scaled() {
+		$attachment = $this->create_image( [ 'file' => 'huge.jpg' ] );
+		$result     = get_timber_image( $attachment, 'full' );
+
+		$image = ' src="' . $this->get_upload_url() . '/huge-scaled.jpg" alt=""';
+
+		$this->assertEquals( $image, $result );
+	}
+
+	/**
+	 * Tests whether a non-scaled version of the image is returned when the
+	 * `original` size is requested.
+	 */
+	public function test_get_timber_image_original_nonscaled() {
+		$attachment = $this->create_image( [ 'file' => 'huge.jpg' ] );
+		$result     = get_timber_image( $attachment, 'original' );
+
+		$image = ' src="' . $this->get_upload_url() . '/huge.jpg" alt=""';
+
+		$this->assertEquals( $image, $result );
+	}
+
+	/**
+	 * Tests whether the 'big_image_size_threshold' works properly and non-scaled version of the
+	 * image is returned.
+	 */
+	public function test_get_timber_image_full_ignored_threshold() {
+		add_filter( 'big_image_size_threshold', '__return_false' );
+
+		$attachment = $this->create_image( [ 'file' => 'huge.jpg' ] );
+		$result     = get_timber_image( $attachment, 'full' );
+
+		$image = ' src="' . $this->get_upload_url() . '/huge.jpg" alt=""';
+
+		$this->assertEquals( $image, $result );
+
+		remove_filter( 'big_image_size_threshold', '__return_false' );
 	}
 
 	public function test_get_timber_image_texts() {
@@ -60,7 +120,7 @@ class TestFunctions extends TimmyUnitTestCase {
 			'alt'    => $alt_text,
 		];
 
-		$this->assertEquals( $result, $attributes );
+		$this->assertEquals( $attributes, $result );
 	}
 
 	public function test_get_timber_image_responsive() {
@@ -70,7 +130,7 @@ class TestFunctions extends TimmyUnitTestCase {
 
 		$image = ' sizes="100vw" srcset="' . $this->get_upload_url() . '/test-560x0-c-default.jpg 560w, ' . $this->get_upload_url() . '/test-1400x0-c-default.jpg 1400w" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7" alt="Burrito Wrap"';
 
-		$this->assertEquals( $result, $image );
+		$this->assertEquals( $image, $result );
 	}
 
 	public function test_get_timber_image_responsive_src() {
@@ -79,7 +139,7 @@ class TestFunctions extends TimmyUnitTestCase {
 
 		$image = ' sizes="100vw" srcset="' . $this->get_upload_url() . '/test-560x0-c-default.jpg 560w, ' . $this->get_upload_url() . '/test-1400x0-c-default.jpg 1400w" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"';
 
-		$this->assertEquals( $result, $image );
+		$this->assertEquals( $image, $result );
 	}
 
 	public function test_get_timber_image_responsive_src_lazy_args() {
@@ -95,14 +155,14 @@ class TestFunctions extends TimmyUnitTestCase {
 			$this->get_upload_url()
 		);
 
-		$this->assertEquals( $result, $image );
+		$this->assertEquals( $image, $result );
 	}
 
 	public function test_get_post_thumbnail() {
 		$attachment = $this->create_image();
 		$result     = get_post_thumbnail( $attachment );
 
-		$this->assertEquals( $result, false );
+		$this->assertEquals( false, $result );
 	}
 
 	public function test_timber_image_full_with_gif() {
@@ -110,6 +170,34 @@ class TestFunctions extends TimmyUnitTestCase {
 		$result     = get_timber_image( $attachment, 'full' );
 
 		$image = ' src="' . $this->get_upload_url() . '/logo-small.gif" alt=""';
+
+		$this->assertEquals( $image, $result );
+	}
+
+	/**
+	 * Tests whether we get the full src of an SVG with size full.
+	 *
+	 * @since 0.14.4
+	 */
+	public function test_timber_image_full_with_svg() {
+		$attachment = $this->create_image( [ 'file' => 'sveegee.svg' ] );
+		$result     = get_timber_image( $attachment, 'full' );
+
+		$image = ' src="' . $this->get_upload_url() . '/sveegee.svg" alt=""';
+
+		$this->assertEquals( $image, $result );
+	}
+
+	/**
+	 * Tests whether we get the full src of an SVG with size large.
+	 *
+	 * @since 0.14.4
+	 */
+	public function test_timber_image_large_with_svg() {
+		$attachment = $this->create_image( [ 'file' => 'sveegee.svg' ] );
+		$result     = get_timber_image( $attachment, 'large' );
+
+		$image = ' src="' . $this->get_upload_url() . '/sveegee.svg" alt=""';
 
 		$this->assertEquals( $image, $result );
 	}